+++ packages/connectivity_status/lib/connectivity_status.dart
@@ -1,4 +1,5 @@
 export 'connectivity_status_platform_interface.dart';
+export 'src/create_connectivity_repository.dart';
 export 'src/data/connectivity_data_source.dart';
 export 'src/data/connectivity_data_source_impl.dart';
 export 'src/data/connectivity_metered_data_source.dart';